+++ NEWS.md
@@ -83,6 +83,8 @@
 
 ### Distributed R
 
+- Fix to `spark_apply()` to retain column types when NAs are present (#1665).
+
 - `spark_apply()` now supports `rlang` anonymous functions. For example,
   `sdf_len(sc, 3) %>% spark_apply(~.x+1)`.